requirements.txt
Flask==2.1.2
google-ads==16.0.0
eventlet==0.33.0
./grpc
